Change reply timeout action and block requesting
If a peer does not respond in REPLY_TIMEOUT_MSEC time, he is marked as
timed-out.
When searching for a block to request, the list of Pieces is cheched
first for unrequested blocks, and if none are found, an attempt is made
to request an already requested block from a timed-out Peer.