Add torrent resuming
After adding a new torrent/shutting down the client, information
needed for resuming the torrent the next time the application is
started. The ResumeInfo class is used to keep this information.
The information is bencoded and saved to the file
AppDataLocation/resume/resume.dat, where AppDataLocation is
ahieved from QStandardPaths
Resuming is initiated from the TorrentManager class.